refactor(): ModalButton, ModalWizard and add(): Wizard
