+++ src/atoms/cta/ModalButton.ts
@@ -1,8 +1,8 @@
-import { ModalWizard } from './ModalWizard';
 import { AnalyticalEvent } from '../../events/AnalyticalEvent';
+import { Modal } from '../../molecules/modal/Modal';
 
 export class ModalButton {
     clickEvent?: AnalyticalEvent;
     label: string;
-    modal: ModalWizard;
+    modal: Modal;
 }
