+++ src/molecules/cta/CtaBlock.ts
@@ -1,7 +1,7 @@
 import { ContentButton } from '../../atoms/cta/ContentButton';
 import { ContentLink } from '../../atoms/cta/ContentLink';
 import { YoutubeLink } from '../../atoms/cta/YoutubeLink';
-import { ModalButton } from '../wizard/ModalButton';
+import { ModalButton } from '../../atoms/cta/ModalButton';
 
 export class CtaBlock {
     ctas: (ContentButton | ContentLink | YoutubeLink | ModalButton)[];
