+++ src/molecules/modal/Modal.ts
@@ -1,0 +1,5 @@
+import { Wizard } from '../wizard/Wizard';
+
+export class Modal {
+    content: Wizard;
+}
